# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Kubefirst API",
        "title": "Kubefirst API",
        "contact": {
            "name": "Kubefirst",
            "email": "help@kubefirst.io"
        },
        "version": "1.0"
    },
    "host": "localhost:port",
    "basePath": "/api/v1",
    "paths": {
...
        "/services/:cluster_name": {
            "get": {
                "description": "Returns a list of services for a cluster",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Returns a list of services for a cluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster name",
                        "name": "cluster_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAPI key\u003e",
                        "description": "API key",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ClusterServiceList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.JSONFailureResponse"
                        }
                    }
                }
            }
        },
        "/services/:cluster_name/:service_name": {
            "post": {
                "description": "Add a gitops catalog application to a cluster as a service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Add a gitops catalog application to a cluster as a service",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster name",
                        "name": "cluster_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Service name to be added",
                        "name": "service_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Service create request in JSON format",
                        "name": "definition",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.GitopsCatalogAppCreateRequest"
                        }
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAPI key\u003e",
                        "description": "API key",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/types.JSONSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.JSONFailureResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a gitops catalog application from a cluster",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Remove a gitops catalog application from a cluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster name",
                        "name": "cluster_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Service name to be removed",
                        "name": "service_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAPI key\u003e",
                        "description": "API key",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/types.JSONSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.JSONFailureResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.ClusterServiceList": {
            "type": "object",
            "properties": {
                "cluster_name": {
                    "type": "string"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Service"
                    }
                }
            }
        },
...
        "types.GitopsCatalogAppCreateRequest": {
            "type": "object",
            "properties": {
                "secret_keys": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.GitopsCatalogAppSecretKey"
                    }
                }
            }
        },
        "types.GitopsCatalogAppSecretKey": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "types.JSONFailureResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "err"
                }
            }
        },
...
        "types.JSONSuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "types.Service": {
            "type": "object",
            "properties": {
                "default": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "links": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}